+++ service/reward.go
@@ -16,7 +16,7 @@ type RewardsRes struct {
 func FetchRewards(page, size int) (res RewardsRes, err error) {
 	err = DB.Model(&model.Reward{}).Order("block_number DESC").Offset((page - 1) * size).Limit(size).
 		Joins("LEFT JOIN collections ON id=LEFT(snft,40)").Select("rewards.*, name AS collection_name").Scan(&res.Rewards).Error
-	res.Total = int64((stats.TotalBlock - 1) * 11)
+	res.Total = (stats.TotalBlock - 1) * 11
 	return
 }
